tools/python_scorpion/pyproject.toml
[project]
name = "python-scorpion"
version = "0.1.0"
description = "Scorpion security toolkit (Python heavy-weight variant)"
readme = "README.md"
authors = [{ name = "Scorpion Team" }]
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "httpx>=0.27.0",
  "dnspython>=2.6.1",
  "cryptography>=43.0.0",
  "uvloop>=0.20.0; platform_system != 'Windows'",
]

[project.scripts]
scorpion = "python_scorpion.cli:app"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
